endTrip now adds points

+++ src/main/java/com/collabothon2022/pedalpower/external/api/google/GoogleDistanceApi.java
@@ -1,5 +1,5 @@
 package com.collabothon2022.pedalpower.external.api.google;
 
 public interface GoogleDistanceApi {
-    double getDistanceInKm(String gpsPointData);
+    int getDistanceInKm(String gpsPointData);
 }
